// client/src/redux/constants.ts
export enum Powers {
    GERMANY = 'GERMANY',
    SOVIET_UNION = 'SOVIET_UNION',
    COMMUNIST_CHINA = 'COMMUNIST_CHINA',
    JAPAN = 'JAPAN',
    UK_WEST = 'UK_WEST',
    UK_EAST = 'UK_EAST',
    ANZAC = 'ANZAC',
    FRANCE = 'FRANCE',
    ITALY = 'ITALY',
    USA = 'USA',
    NATIONALIST_CHINA = 'NATIONALIST_CHINA',
}

export enum Seasons {
    WINTER = 'WINTER',
    SUMMER = 'SUMMER',
}

export const START_SEASON = Seasons.SUMMER;
export const START_YEAR = 1936;

export const ORDER = [
    Powers.GERMANY,
    Powers.SOVIET_UNION,
    Powers.COMMUNIST_CHINA,
    Powers.JAPAN,
    Powers.UK_WEST,
    Powers.UK_EAST,
    Powers.ANZAC,
    Powers.FRANCE,
    Powers.ITALY,
    Powers.USA,
    Powers.NATIONALIST_CHINA,
];